for mentioning the IKEv2 COOKIE system
The ACS should be robust enough to resist attacks on itself.

RPKI authentication is not mentioned here. On the one hand, it is high cost; on the other hand, RPKI is not an authentication but an authorization mechanism. See RFC 6480, Section 2.1.  Role in the Overall Architecture. <https://www.rfc-editor.org/rfc/rfc6480.html>